+++ themes/theme-rui/src/components/Link.styles.ts
@@ -1,5 +1,16 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const Link: ThemeComponent<'Link'> = cva(
-  'font-medium underline aria-[disabled]:cursor-not-allowed underline-offset-4'
+  ' text-link font-normal aria-[disabled]:cursor-not-allowed',
+  {
+    variants: {
+      variant: {
+        default: '',
+        secondary: 'font-medium text-foreground underline hover:no-underline',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+    },
+  }
 );